refactor(coupon): 적용 가능한 쿠폰 먼저 보여주는 로직 list 로 이동
