+++ src/pages/CouponModal/CouponModal.tsx
@@ -6,11 +6,9 @@ import DeleteButton from "@/assets/delete-icon.svg?react";
 import { Coupon } from "@/types/coupon";
 import Button from "@/components/_common/Button/Button";
 import { theme } from "@/styles/theme";
+import CouponList from "./components/CouponList";
 import { useRecoilValue } from "recoil";
 import { couponListSelector } from "@/recoil/coupons";
-import useCouponApplicabilityChecker from "@/hooks/coupon/useCouponApplicabilityChecker";
-import { totalOrderPriceSelector } from "@/recoil/orderInformation";
-import CouponList from "./components/CouponList";
 
 export type CouponWithApplicablity = { coupon: Coupon; applicability: boolean };
 
@@ -22,31 +20,6 @@ const CouponModal = ({
   onCloseModal: () => void;
 }) => {
   const couponList = useRecoilValue(couponListSelector);
-  const totalItemsPrice = useRecoilValue(totalOrderPriceSelector);
-  const { isCouponApplicable } = useCouponApplicabilityChecker();
-
-  const applicableCoupons: CouponWithApplicablity[] = [];
-  const nonApplicableCoupons: CouponWithApplicablity[] = [];
-
-  couponList.forEach((coupon) => {
-    if (
-      isCouponApplicable({
-        coupon: coupon,
-        price: totalItemsPrice,
-        time: new Date(),
-      })
-    ) {
-      applicableCoupons.push({ coupon: coupon, applicability: true });
-    } else {
-      nonApplicableCoupons.push({ coupon: coupon, applicability: false });
-    }
-  });
-
-  const sortedCouponsWithApplicability = [
-    ...applicableCoupons,
-    ...nonApplicableCoupons,
-  ];
-
   return (
     <Modal
       isOpen={isOpen}
@@ -66,7 +39,7 @@ const CouponModal = ({
           <DeleteButton />
         </Modal.CloseIcon>
 
-        <CouponList couponsWithApplicability={sortedCouponsWithApplicability} />
+        <CouponList couponList={couponList} />
         <Button
           onClick={() => {}}
           style={{
