+++ src/pages/CouponModal/components/CouponList.tsx
@@ -1,15 +1,34 @@
-import { CouponWithApplicablity } from "../CouponModal";
+import useCouponApplicabilityChecker from "@/hooks/coupon/useCouponApplicabilityChecker";
 import CouponItem from "./Coupon";
+import { CouponWithApplicablity } from "../CouponModal";
+import { Coupon } from "@/types/coupon";
+import { useRecoilValue } from "recoil";
+import { totalOrderPriceSelector } from "@/recoil/orderInformation";
 
-const CouponList = ({
-  couponsWithApplicability,
-}: {
-  couponsWithApplicability: CouponWithApplicablity[];
-}) => {
+const CouponList = ({ couponList }: { couponList: Coupon[] }) => {
+  const totalItemsPrice = useRecoilValue(totalOrderPriceSelector);
+  const { isCouponApplicable } = useCouponApplicabilityChecker();
+
+  const applicableCoupons: CouponWithApplicablity[] = [];
+  const nonApplicableCoupons: CouponWithApplicablity[] = [];
+
+  couponList.forEach((coupon) => {
+    if (isCouponApplicable(coupon, totalItemsPrice, new Date())) {
+      applicableCoupons.push({ coupon: coupon, applicability: true });
+    } else {
+      nonApplicableCoupons.push({ coupon: coupon, applicability: false });
+    }
+  });
+
+  const sortedCouponsWithApplicability = [
+    ...applicableCoupons,
+    ...nonApplicableCoupons,
+  ];
+
   return (
     <>
-      {couponsWithApplicability.map((oneCoupon) => {
-        const { coupon, applicability } = oneCoupon;
+      {sortedCouponsWithApplicability.map((couponWithApplicability) => {
+        const { coupon, applicability } = couponWithApplicability;
         return (
           <CouponItem
             key={coupon.id}
